python/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.18
PyMySQL==1.0.2
openai==0.28
cryptography==39.0.1
flask-cors==3.2.1
flask-executor==1.5.2
rapidfuzz==2.7.1
python-dotenv==1.0.1